+++ internal/storage/postgresql/migrations/177_add_changelog_v5_1_1.down.sql
@@ -1,0 +1,1 @@
+DELETE FROM changelogs WHERE version = '5.1.1';